LayoutTests/imported/w3c/web-platform-tests/css/selectors/has-focus-display-change-expected.txt

A

FAIL :has(:focus) doesn't cause focusability to be lost assert_equals: Focus can move to #second expected Element node <button id="second">B</button> but got Element node <body><div id="container">
  <button id="first">A</button...
